Remove unneeded casts.
git-svn-id: svn+ssh://src.earth.threerings.net/nenya/trunk@397 ed5b42cb-e716-0410-a449-f6a68f950b19

+++ src/java/com/threerings/miso/util/MisoSceneMetrics.java
@@ -87,11 +87,11 @@ public class MisoSceneMetrics
         slopeY = -slopeX;
 
         // calculate the edge length separating each fine coordinate
-        finelen = tilelen / (float)finegran;
+        finelen = tilelen / finegran;
 
         // calculate the fine-coordinate x-axis line
         fineSlopeX = (float)tilehei / (float)tilewid;
-        fineBX = -(fineSlopeX * (float)tilehwid);
+        fineBX = -(fineSlopeX * tilehwid);
         fineSlopeY = -fineSlopeX;
 
         // calculate the fine coordinate dimensions
